requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.14.0
mccabe==0.7.0
numpy==1.26.4
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
shapely==2.0.3
sqlparse==0.4.4
typing_extensions==4.11.0
